+++ packages/polaris-viz/src/components/LineChart/Chart.tsx
@@ -82,6 +82,7 @@ export interface ChartProps {
     chart?: (props: LineChartSlotProps) => JSX.Element;
   };
   theme?: string;
+  enableHideLegendOverflow?: boolean;
 }
 
 export function Chart({
@@ -97,6 +98,7 @@ export function Chart({
   theme = DEFAULT_THEME_NAME,
   xAxisOptions,
   yAxisOptions,
+  enableHideLegendOverflow = false,
 }: ChartProps) {
   useColorVisionEvents({
     enabled: data.length > 1,
@@ -433,7 +435,7 @@ export function Chart({
           renderHiddenLegendLabel={renderHiddenLegendLabel}
           width={width}
           dimensions={dimensions}
-          enableHideOverflow
+          enableHideOverflow={enableHideLegendOverflow}
         />
       )}
     </Fragment>