+++ packages/polaris-viz/src/components/LineChart/LineChart.tsx
@@ -47,6 +47,7 @@ export type LineChartProps = {
   slots?: {
     chart?: (props: LineChartSlotProps) => JSX.Element;
   };
+  enableHideLegendOverflow?: boolean;
 } & ChartProps;
 
 export function LineChart(props: LineChartProps) {
@@ -69,6 +70,7 @@ export function LineChart(props: LineChartProps) {
     tooltipOptions,
     xAxisOptions,
     yAxisOptions,
+    enableHideLegendOverflow = false,
   } = {
     ...DEFAULT_CHART_PROPS,
     ...props,
@@ -119,6 +121,7 @@ export function LineChart(props: LineChartProps) {
             xAxisOptions={xAxisOptionsWithDefaults}
             yAxisOptions={yAxisOptionsWithDefaults}
             slots={props.slots}
+            enableHideLegendOverflow={enableHideLegendOverflow}
           />
         )}
       </ChartContainer>