+++ packages/polaris-viz/src/components/StackedAreaChart/Chart.tsx
@@ -80,6 +80,8 @@ export interface Props {
   yAxisOptions: Required<YAxisOptions>;
   dimensions?: Dimensions;
   renderLegendContent?: RenderLegendContent;
+  enableHideLegendOverflow?: boolean;
+  renderHiddenLegendLabel?: (count: number) => string;
 }
 
 export function Chart({
@@ -92,6 +94,8 @@ export function Chart({
   showLegend,
   theme,
   yAxisOptions,
+  enableHideLegendOverflow = false,
+  renderHiddenLegendLabel,
 }: Props) {
   useColorVisionEvents({enabled: data.length > 1});
 
@@ -398,8 +402,9 @@ export function Chart({
           onDimensionChange={setLegendDimensions}
           renderLegendContent={renderLegendContent}
           width={width}
-          enableHideOverflow
+          enableHideOverflow={enableHideLegendOverflow}
           dimensions={chartBounds}
+          renderHiddenLegendLabel={renderHiddenLegendLabel}
         />
       )}
     </ChartElements.Div>