+++ packages/polaris-viz/src/components/VerticalBarChart/VerticalBarChart.tsx
@@ -28,6 +28,8 @@ export interface VerticalBarChartProps {
   emptyStateText?: string;
   renderLegendContent?: RenderLegendContent;
   type?: ChartType;
+  enableHideLegendOverflow?: boolean;
+  renderHiddenLegendLabel?: (count: number) => string;
 }
 
 export function VerticalBarChart({
@@ -41,6 +43,8 @@ export function VerticalBarChart({
   type = 'default',
   xAxisOptions,
   yAxisOptions,
+  enableHideLegendOverflow = false,
+  renderHiddenLegendLabel,
 }: VerticalBarChartProps) {
   const selectedTheme = useTheme();
   const seriesColors = useThemeSeriesColors(data, selectedTheme);
@@ -62,6 +66,8 @@ export function VerticalBarChart({
       type={type}
       xAxisOptions={xAxisOptions}
       yAxisOptions={yAxisOptions}
+      enableHideLegendOverflow={enableHideLegendOverflow}
+      renderHiddenLegendLabel={renderHiddenLegendLabel}
     />
   );
 }